Реализовал логику работы с автомоечными средствами в боксе

+++ src/main/java/ru/pin120/carwashAPI/controllers/CategoriesOfSuppliesController.java
@@ -73,13 +73,11 @@ public class CategoriesOfSuppliesController {
 
             CategoryOfSupplies categoryOfSupplies = categoryOfSuppliesOptional.get();
 
-            // написать условие для удаления
-//            if(!category.getServices().isEmpty()){
-//                return new ResponseEntity<>(String.format("Нельзя удалить категорию %s, так как у неё есть услуги", category.getCatName()), HttpStatus.BAD_REQUEST);
-//            }else{
-//                categoryOfServicesService.delete(category.getCatName());
-//            }
-            categoriesOfSuppliesService.delete(categoryOfSupplies);
+            if(!categoryOfSupplies.getSupplies().isEmpty()){
+                return new ResponseEntity<>(String.format("Нельзя удалить категорию %s, так как она указана в автомоечном средстве", categoryOfSupplies.getCSupName()), HttpStatus.BAD_REQUEST);
+            }else{
+                categoriesOfSuppliesService.delete(categoryOfSupplies);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
